update final 2

+++ S4D/s4d_server/DTOs/Request/UserRequestDTO.cs
@@ -7,7 +7,7 @@
         public string? Password { get; set; }
         public string? Email { get; set; }
         public string? Role { get; set; }
-        public ulong Status { get; set; }
+        public byte Status { get; set; }
         public string? Address { get; set; } 
         public string? PhoneNumber { get; set; }
     }
